Move maybe_add_argument to functional.py
